specify data types in test

+++ test/unit/ard/collection/test_optiwindnet.py
@@ -26,10 +26,10 @@ class TestOptiWindNetCollection:
         self.farm_spec = {}
         self.farm_spec["xD_farm"], self.farm_spec["yD_farm"] = [
             7 * v.flatten()
-            for v in np.meshgrid(np.linspace(-2, 2, 5), np.linspace(-2, 2, 5))
+            for v in np.meshgrid(np.linspace(-2, 2, 5, dtype=int), np.linspace(-2, 2, 5, dtype=int))
         ]
-        self.farm_spec["x_substations"] = np.array([-500.0, 500.0])
-        self.farm_spec["y_substations"] = np.array([-500.0, 500.0])
+        self.farm_spec["x_substations"] = np.array([-500.0, 500.0], dtype=np.float64)
+        self.farm_spec["y_substations"] = np.array([-500.0, 500.0], dtype=np.float64)
 
         # specify the configuration/specification files to use
         filename_turbine_spec = (
